Migrate blog to feature module (PSY-103)
Move blog components, utilities, and types into features/blog/
with co-located structure. Merge blog and mix types into single
types.ts. Update all imports.

Closes PSY-103
