+++ frontend/app/page.tsx
@@ -1,9 +1,6 @@
 import Link from 'next/link'
 import { HomeShowList } from '@/components/shows'
-import { getBlogSlugs, getBlogPost } from '@/lib/blog'
-import { getAllMixes } from '@/lib/mixes'
-import { MDXContent } from '@/components/blog/mdx-content'
-import { SoundCloud } from '@/components/blog/soundcloud-embed'
+import { getBlogSlugs, getBlogPost, getAllMixes, MDXContent, SoundCloud } from '@/features/blog'
 import { JsonLd } from '@/components/seo/JsonLd'
 import { generateWebSiteSchema } from '@/lib/seo/jsonld'
 import { formatContentDate } from '@/lib/utils/formatters'
