+++ frontend/features/blog/components/index.ts
@@ -1,3 +1,3 @@
 export { Bandcamp } from './bandcamp-embed'
 export { SoundCloud } from './soundcloud-embed'
-
+export { MDXContent } from './mdx-content'
